# retouches pill sprite codes a bit, and adds new pill options + patch shape options in chem master