--- a/code/__DEFINES/reagents.dm
+++ b/code/__DEFINES/reagents.dm
@@ -58,10 +58,10 @@
 
 /* <pill sprite size standard>
 	 Since sprite asset code crops the pill image, you are required to make a pill image within [10,24,22,9] squared area.
-	 There is a dummy image that you can recognise the size of a cropped pill image in 'chemical.dmi'
+	 There is a dummy image that you can recognise the size of a cropped pill image in 'pills.dmi'
 	 The black line counts, so you can use that area for your sprite as well. */
 
-// pill shapes - check 'chemical.dmi' for the shape
+// pill shapes - check 'pills.dmi' for the shape
 #define PILL_SHAPE_LIST list(\
 	"pill_shape_capsule_purple_pink",\
 	"pill_shape_capsule_bloodred",\