… code/modules/plumbing/plumbers/pill_press.dm
… code/modules/reagents/chemistry/machinery/chem_master.dm
… code/modules/reagents/reagent_containers/patch.dm
… code/modules/reagents/reagent_containers/pill.dm
… icons/UI_Icons/Pills/pill1.png
… icons/UI_Icons/Pills/pill10.png
… icons/UI_Icons/Pills/pill11.png
… icons/UI_Icons/Pills/pill12.png
… icons/UI_Icons/Pills/pill13.png
… icons/UI_Icons/Pills/pill14.png
… icons/UI_Icons/Pills/pill15.png
… icons/UI_Icons/Pills/pill16.png
… icons/UI_Icons/Pills/pill17.png
… icons/UI_Icons/Pills/pill18.png
… icons/UI_Icons/Pills/pill19.png
… icons/UI_Icons/Pills/pill2.png
… icons/UI_Icons/Pills/pill20.png
… icons/UI_Icons/Pills/pill21.png
… icons/UI_Icons/Pills/pill22.png
… icons/UI_Icons/Pills/pill3.png
… icons/UI_Icons/Pills/pill4.png
… icons/UI_Icons/Pills/pill5.png
… icons/UI_Icons/Pills/pill6.png
… icons/UI_Icons/Pills/pill7.png
… icons/UI_Icons/Pills/pill8.png
… icons/UI_Icons/Pills/pill9.png
… icons/obj/chemical.dmi
… icons/obj/medicine_containers.dmi
… icons/obj/pills.dmi
… tgui/packages/tgui/interfaces/ChemPress.js