<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull request process. -->

## About The Pull Request
* changes the pill asset code
* adds more pill options (adds new pill sprites too)

<!-- Describe The Pull Request. Please be sure every change is documented or this can delay review and even discourage maintainers from merging your PR! -->

## Why It's Good For The Game
`pill1 2 3 4 5` is awful to recognise.
the best is using greyscale, but It'll be replaceable later...
also, touched some bad code
<!-- Please add a short description of why you think these changes would benefit the game. If you can't justify it in words, it might not be worth adding.
-->

## Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots&Videos</summary>

![image](https://user-images.githubusercontent.com/87972842/212142844-781b6947-e5ef-4783-839b-041a703b6101.png)

pills.dmi
images are relocated, and added new pill sprites

![image](https://user-images.githubusercontent.com/87972842/212142965-c001da40-3eb8-4348-8472-d2baebcd5832.png)

![image](https://user-images.githubusercontent.com/87972842/211348138-43fa3e54-8c90-4d4a-b60e-25c587617e67.png)

![image](https://user-images.githubusercontent.com/87972842/212143920-cf83f844-6f29-4e26-b232-d4335e3548e0.png)


new pill options

![image](https://user-images.githubusercontent.com/87972842/211348195-e1939488-40ce-471d-86cd-67de831d90bc.png)

![image](https://user-images.githubusercontent.com/87972842/212144051-725830a2-68a2-4691-99a9-172e0613d0cf.png)


random pill option works well

----------

![image](https://user-images.githubusercontent.com/87972842/211355073-e975db20-943d-4d4e-abd1-46da61114aa0.png)

this is an example image that you can recognise a proper size of a pill sprite

</details>

## Changelog
:cl:
code: fixed some code that didn't use random pill shape correctly
add: "medicine_containers.dmi" that contains pill sprites, pill bottle sprites and patch sprites
del: weirdly implemented pill image folder
code: medicine_containers image asset that is used by chemmaster tgui. old pill asset list is removed.
tweak: chem master will now use more wider pill & patch options
imagedel: removed unused pill image (pill0) which is actually a duplication of a pill image(pill18). also arranged pill image order.
imageadd: angry pill, sad pill, and question pill (random pick one)... new pill colours - pink, indigo, red, orange, etc... and pill example dummy image.
imageadd: blank patch sprites
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put your name to the right of the first :cl: if you want to overwrite your GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the icon ingame) and delete the unneeded ones. Despite some of the tags, changelogs should generally represent how a player might be affected by the changes rather than a summary of the PR's contents. -->