@EvilDragonfiend 2023-01-09
![image](https://user-images.githubusercontent.com/87972842/211355717-2f234ce7-236c-4102-8fa3-64e9e7661aff.png)

this doesn't happen in my local.

```
Logging in...connected
Logged in as EvilDragonfiend.
Tue Jan 10 01:18:24 2023
World opened on network port 2555.
Welcome BYOND! (5.0 Public Version 514.1583)
813 global variables
World loaded at 01:18:31!
Loading config file config.txt...
Loading config file game_options.txt...
Failed to validate setting "ASSIMILATION 0" for probability
Failed to validate setting "ASSIMILATION" for continuous
Failed to validate setting "DOUBLE_AGENTS 25" for min_pop
Failed to validate setting "DOUBLE_AGENTS -1" for max_pop
Failed to validate setting "ASSIMILATION 0" for min_pop
Failed to validate setting "ASSIMILATION -1" for max_pop
Loading config file gameplay_tweaks.txt...
Loading config file dbconfig.txt...
Loading config file comms.txt...
Loading config file antag_rep.txt...
Loading config file resources.txt...
Loading config file interviews.txt...
Failed to validate setting "Why have you joined the server today?" for interview_questions
Failed to validate setting "Have you played space-station 13 before? If so, on what servers?" for interview_questions
Failed to validate setting "Do you know anybody on the server today? If so, who?" for interview_questions
Failed to validate setting "Do you have any additional comments?" for interview_questions
Loading config file performance.txt...
Loading config file maps.txt...
Loading config file topic_rate_limit_whitelist.txt...
Loading config file protected_cids.json...
Loading config file ooc_filter.txt...
Loading config file in_character_filter.txt...
Running /tg/ revision: 2023-01-09
origin/master: d7f9b13e28ed32e75f9fc0a9c42cee4e46807da2
HEAD: 699f150f02a316592f91c459ed47b9f257fd7030

Webclient users can visit this address to play:
http://www.byond.com/play/220.76.136.179:2555 

Tue Jan 10 01:18:32 2023

Or this code can be embedded:
<iframe src="http://www.byond.com/play/embed/220.76.136.179:2555" width=640 height=480></iframe>

Initialized Fail2Topic subsystem within 0 seconds!
Initialized Profiler subsystem within 0 seconds!
Initialized Title Screen subsystem within 0 seconds!
Initialized Database subsystem within 0 seconds!
Initialized Blackbox subsystem within 0 seconds!
Initialized Server Tasks subsystem within 0 seconds!
Initialized Input subsystem within 0 seconds!
Initialized Topic subsystem within 0 seconds!
Initialized Sound Channels subsystem within 0 seconds!
Initialized Instruments subsystem within 0 seconds!
Initialized Greyscale subsystem within 0.2 seconds!
Initialized Vis contents overlays subsystem within 0 seconds!
Initialized Achievements subsystem within 0 seconds!
Initialized Materials subsystem within 0 seconds!
## WARNING: Invalid research design with ID sleepy detected in node Weapon Development Technology[weaponry] removed. in code/controllers/subsystem/research.dm at line 190 src: Research usr: .
Initialized Research subsystem within 0 seconds!
Initialized Orbits subsystem within 0 seconds!
Initialized Station subsystem within 0 seconds!
Initialized Events subsystem within 0 seconds!
Initialized Jobs subsystem within 0 seconds!
Initialized Quirks subsystem within 0 seconds!
Initialized AI movement subsystem within 0 seconds!
Master mode is 'secret_extended'
Initialized Ticker subsystem within 0 seconds!
Initialized AI Behavior Ticker subsystem within 0 seconds!
Initialized AI Controller Ticker subsystem within 0 seconds!
Loading Runtime Station...
Loaded Station in 0.1s!
Loaded Random Rooms in 0s!
Initialized Mapping subsystem within 0.4 seconds!
Initialized Time Tracking subsystem within 0 seconds!
Initialized Networks subsystem within 0 seconds!
Initialized Economy subsystem within 0 seconds!
## WARNING: docking port 'pod_lavaland1' could not be randomly placed in /area/lavaland/surface/outdoors: of 0 turfs, none were suitable. in code/modules/shuttle/emergency.dm at line 660 src: the lavaland usr: .
Initialized Atoms subsystem within 3.3 seconds!
Initialized Language subsystem within 0 seconds!
Initialized Machines subsystem within 0 seconds!
Initialized Atmos Adjacency subsystem within 0 seconds!
Initialized Autotransfer Vote subsystem within 0 seconds!
Initialized Circuit Components subsystem within 0 seconds!
Initialized Disease subsystem within 0 seconds!
Initialized Metrics subsystem within 0 seconds!
Initialized Night Shift subsystem within 0 seconds!
Initialized Parallax subsystem within 0 seconds!
Initialized Traumas subsystem within 0 seconds!
Initialized Weather subsystem within 0 seconds!
Initialized Atmospherics subsystem within 0.5 seconds!
Loaded 0 engraved messages on map Runtime Station
Initialized Persistence subsystem within 0 seconds!
## WARNING: design /datum/design/leftarm with icon 'icons/mob/human_parts_greyscale.dmi' missing state 'default_human_l_arm'
## WARNING: design /datum/design/rightarm with icon 'icons/mob/human_parts_greyscale.dmi' missing state 'default_human_r_arm'
## WARNING: design /datum/design/leftleg with icon 'icons/mob/human_parts_greyscale.dmi' missing state 'default_human_l_leg'
## WARNING: design /datum/design/rightleg with icon 'icons/mob/human_parts_greyscale.dmi' missing state 'default_human_r_leg'
## WARNING: design /datum/design/borg_chest with icon '' missing state 'borg_chest'
## WARNING: design /datum/design/borg_head with icon '' missing state 'borg_head'
## WARNING: design /datum/design/borg_l_arm with icon '' missing state 'borg_l_arm'
## WARNING: design /datum/design/borg_r_arm with icon '' missing state 'borg_r_arm'
## WARNING: design /datum/design/borg_l_leg with icon '' missing state 'borg_l_leg'
## WARNING: design /datum/design/borg_r_leg with icon '' missing state 'borg_r_leg'
Initialized Assets subsystem within 2.4 seconds!
Initialized Icon Smoothing subsystem within 0.1 seconds!
Initialized Overlay subsystem within 0 seconds!
Initialized XKeyScore subsystem within 0 seconds!
Initialized PRISM subsystem within 0 seconds!
Initialized Lighting subsystem within 0.4 seconds!
## WARNING: No /obj/docking_port/mobile/arrivals placed on the map! in code/controllers/subsystem/shuttle.dm at line 83 src: Shuttle usr: .
## WARNING: No /obj/docking_port/mobile/emergency placed on the map! in code/controllers/subsystem/shuttle.dm at line 85 src: Shuttle usr: .
## WARNING: No /obj/docking_port/mobile/emergency/backup placed on the map! in code/controllers/subsystem/shuttle.dm at line 87 src: Shuttle usr: .
## WARNING: No /obj/docking_port/mobile/supply placed on the map! in code/controllers/subsystem/shuttle.dm at line 89 src: Shuttle usr: .
Initialized Shuttle subsystem within 0.1 seconds!
Initialized Minor Mapping subsystem within 0.1 seconds!
Initialized Pathfinder subsystem within 0 seconds!
Initialized Chat subsystem within 0 seconds!
Initializations complete within 7.5 seconds!
```