@EvilDragonfiend 2023-01-10
issue solved. the issue was `"filename.dmi"` which had to be `'filename.dmi'`