@itsmeow 2023-01-11
My bad, I see you converted it to a DMI. It still should use a list though.

Also merging it with chemical.dmi is dumb because you have to crop the pills, just put it in its own DMI where you can have the size of the images make sense.