@EvilDragonfiend 2023-01-12
it is majorly changed now.
* no defines
* more pills