@EvilDragonfiend 2023-01-22
Todo:
* glob will be better than define list since it's often used
* patches