@EvilDragonfiend 2023-01-24
maybe sprite boxes seem to be too big. I think I should make it a bit smaller.