@EvilDragonfiend 2023-01-26
![image](https://user-images.githubusercontent.com/87972842/214742655-e3da5f0b-dd68-47eb-98a2-12b2e1c83a87.png)

now it has less margin.
I think it's ready.